Incorporated use of matfile function to read in files faster
Sometimes we store multiple trace structures in the same .mat file. Previously, the entire file had to be read into memory in order to open any one of the trace structures within it. However, I recently learned that the "matfile" function allows you to only read into memory the part of a .mat file that you need, so I use that now to speed things up. I tested this, and it works!